common/src/test/java/com/salesforce/ouroboros/util/rate/controllers/RateLimiterTest.java
/**
 * Copyright (c) 2011, salesforce.com, inc.
 * All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without modification, are permitted provided
 * that the following conditions are met:
 *
 *    Redistributions of source code must retain the above copyright notice, this list of conditions and the
 *    following disclaimer.
 *
 *    Redistributions in binary form must reproduce the above copyright notice, this list of conditions and
 *    the following disclaimer in the documentation and/or other materials provided with the distribution.
 *
 *    Neither the name of salesforce.com, inc. nor the names of its contributors may be used to endorse or
 *    promote products derived from this software without specific prior written permission.
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED
 * WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
 * PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
 * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
 * TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
 * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
 * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 * POSSIBILITY OF SUCH DAMAGE.
 */
package com.salesforce.ouroboros.util.rate.controllers;

import static junit.framework.Assert.assertEquals;
import static junit.framework.Assert.assertFalse;
import static junit.framework.Assert.assertTrue;

import org.junit.Test;

/**
 * 
 * @author hhildebrand
 * 
 */
public class RateLimiterTest {

    @Test
    public void testLimiter() throws InterruptedException {
        RateLimiter limiter = new RateLimiter(10.0, 10, 0);
        assertTrue(limiter.accept());
        assertEquals(9, limiter.getCurrentTokens());
        for (int i = 0; i < 9; i++) {
            assertTrue(limiter.accept());
        }
        assertFalse(limiter.accept());
        // wait for a token to regenerate
        Thread.sleep(100);
        assertTrue(limiter.accept());
        // clear any remaining tokens 
        for (int i = 0; i < limiter.getCurrentTokens(); i++) {
            assertTrue(limiter.accept());
        }
        assertEquals(0, limiter.getCurrentTokens());

        limiter.setTargetRate(2);
        Thread.sleep(500);
        // Only one token should regenerate
        assertTrue(limiter.accept());
        assertFalse(limiter.accept());
    }
}
